+++ src/rds.h
@@ -37,13 +37,13 @@
 
 #define MAX_AF 25
 
-#define DI_STEREO 1 // 1 - Stereo
-#define DI_AH 2 // 2 - Artificial Head
-#define DI_COMPRESSED 4 // 4 - Compressed
-#define DI_DPTY 8 // 8 - Dynamic PTY
+#define DI_STEREO	1 // 1 - Stereo
+#define DI_AH		2 // 2 - Artificial Head
+#define DI_COMPRESSED	4 // 4 - Compressed
+#define DI_DPTY		8 // 8 - Dynamic PTY
 
 extern int init_rds_encoder(uint16_t pi, char *ps, char *rt, int pty,
-                             int tp, int *af_array, char *ptyn);
+                             int tp, uint8_t *af_array, char *ptyn);
 
 extern float get_rds_sample();
 
@@ -56,7 +56,7 @@ extern void set_rds_rtp_tags(int type_1, int start_1, int len_1,
 extern void set_rds_ta(int ta);
 extern void set_rds_pty(int pty);
 extern void set_rds_ptyn(char *ptyn, int enable);
-extern void set_rds_af(int *af_array);
+extern void set_rds_af(uint8_t *af_array);
 extern void set_rds_tp(int tp);
 extern void set_rds_ms(int ms);
 extern void set_rds_ab(int ab);
